Moved `BasicTimeProfiler` and `HandlersTimeProfiler` to core (#2132)
* restore time_profilers

* moved to core

* updated docs

* edit version in docs

* rearragne imports ignore isort

* add isort skip

* update docs

* added versionadded

* updates

* move tests + edit import
