+++ docs/source/handlers.rst
@@ -17,7 +17,8 @@ Complete list of handlers
     early_stopping.EarlyStopping
     lr_finder.FastaiLRFinder
     terminate_on_nan.TerminateOnNan
-    TimeLimit
+    time_profilers.BasicTimeProfiler
+    time_profilers.HandlersTimeProfiler
     timing.Timer
     global_step_from_engine
     stores.EpochOutputStore
