+++ tests/ignite/handlers/test_time_profilers.py
